prepare support for special columns
